+++ prechi/src/prechi_partition.h
@@ -18,10 +18,7 @@ void prechi_partition_join(PrechiPartition *part, int offset);
 
 int prechi_partition_sorted_offset(PrechiPartition *part, int offset);
 int prechi_partition_value(PrechiPartition *part, int offset);
-int prechi_partition_boundary_count(PrechiPartition *part);
 int prechi_partition_minimum_count(PrechiPartition *part);
-const int *prechi_partition_counts(PrechiPartition *part);
-const int *prechi_partition_spans(PrechiPartition *part);
 float prechi_partition_mean(PrechiPartition *part);
 float prechi_partition_variance(PrechiPartition *part, float mean);
