+++ prechi/test/Makefile.am
@@ -12,7 +12,6 @@ AM_LDFLAGS = -module -rpath $(libdir) -avoid-version -no-undefined
 test_prechi_la_SOURCES = \
   array_init.h array_init.c \
   test_partition.h test_partition.c \
-  test_partition_access.c \
   test_partition_create.c \
   test_partition_join.c \
   test_partition_mean.c \
